requirements.txt
requests
PyYaml
cryptography